+++ redshift_sqlalchemy/dialect.py
@@ -1,4 +1,8 @@
 import pkg_resources
+import re
+import numbers
+
+import sqlalchemy as sa
 from sqlalchemy import schema, util, exc
 from sqlalchemy.dialects.postgresql.base import PGDDLCompiler, PGCompiler
 from sqlalchemy.dialects.postgresql.psycopg2 import PGDialect_psycopg2
@@ -298,7 +302,6 @@ class UnloadFromSelect(Executable, ClauseElement):
 @compiles(UnloadFromSelect)
 def visit_unload_from_select(element, compiler, **kw):
     ''' Returns the actual sql query for the UnloadFromSelect class
-
     '''
     return """
            UNLOAD ('%(query)s') TO '%(unload_location)s'
@@ -320,67 +323,158 @@ def visit_unload_from_select(element, compiler, **kw):
             'parallel': element.options.get('parallel', 'ON')}
 
 
-class CopyCommand(Executable, ClauseElement):
-    ''' Prepares a RedShift COPY statement
-    '''
-    def __init__(self, schema_name, table_name, data_location, access_key, secret_key, session_token='', options={}):
-        ''' Initializes a CopyCommand instance
+# At the time of this implementation, no specification for a session token was
+# found. After looking at a few session tokens they appear to be the same as
+# the aws_secret_access_key pattern, but much longer. An example token can be
+# found here: http://docs.aws.amazon.com/STS/latest/APIReference/API_GetSessionToken.html
+# The regexs for access keys can be found here: http://blogs.aws.amazon.com/security/blog/tag/key+rotation
+creds_rx = re.compile(r"""
+    ^aws_access_key_id=[A-Z0-9]{20};
+    aws_secret_access_key=[A-Za-z0-9/+=]{40}
+    (?:;token=[A-Za-z0-9/+=]+)?$
+""", re.VERBOSE)
 
-        Args:
-            self: An instance of CopyCommand
-            schema_name - Schema associated with the table_name
-            table_name: The table to copy the data into
-            data_location The Amazon S3 location from where to copy - or a manifest file if 'manifest' option is used
-            access_key - AWS Access Key (required)
-            secret_key - AWS Secret Key (required)
-            session_token - AWS STS Session Token (optional)
-            options - Set of optional parameters to modify the COPY sql
-                delimiter - File delimiter; defaults to ','
-                ignore_header - Integer value of number of lines to skip at the start of each file
-                null - Optional string value denoting what to interpret as a NULL value from the file
-                manifest - Boolean value denoting whether data_location is a manifest file; defaults to False
-                empty_as_null - Boolean value denoting whether to load VARCHAR fields with
-                                empty values as NULL instead of empty string; defaults to True
-                blanks_as_null - Boolean value denoting whether to load VARCHAR fields with
-                                 whitespace only values as NULL instead of whitespace; defaults to True
-        '''
-        self.schema_name = schema_name
-        self.table_name = table_name
+
+class CopyCommand(Executable, ClauseElement):
+    """
+    Prepares a Redshift COPY statement.
+
+    Parameters
+    ----------
+    table : sqlalchemy.Table
+        The table to copy data into
+    data_location : str
+        The Amazon S3 location from where to copy, or a manifest file if
+        the `manifest` option is used
+    access_key : str
+    secret_key : str
+    session_token : str, optional
+    delimiter : File delimiter, optional
+        defaults to ','
+    ignore_header : int, optional
+        Integer value of number of lines to skip at the start of each file
+    dangerous_null_delimiter : str, optional
+        Optional string value denoting what to interpret as a NULL value from
+        the file. Note that this parameter *is not properly quoted* due to a
+        difference between redshift's and postgres's COPY commands
+        interpretation of strings. For example, null bytes must be passed to
+        redshift's ``NULL`` verbatim as ``'\\0'`` whereas postgres's ``NULL``
+        accepts ``'\\x00'``.
+    manifest : bool, optional
+        Boolean value denoting whether data_location is a manifest file.
+    empty_as_null : bool, optional
+        Boolean value denoting whether to load VARCHAR fields with empty
+        values as NULL instead of empty string
+    blanks_as_null : bool, optional
+        Boolean value denoting whether to load VARCHAR fields with whitespace
+        only values as NULL instead of whitespace
+    format : str, optional
+        CSV, JSON, or AVRO. Indicates the type of file to copy from.
+    compression : str, optional
+        GZIP, LZOP, indicates the type of compression of the file to copy
+    """
+    formats = ['CSV', 'JSON', 'AVRO']
+    compression_types = ['GZIP', 'LZOP']
+
+    def __init__(self, table, data_location, access_key_id, secret_access_key,
+                 session_token=None, delimiter=',', ignore_header=0,
+                 dangerous_null_delimiter=None, manifest=False,
+                 empty_as_null=True,
+                 blanks_as_null=True, format='CSV', compression=None):
+
+        credentials = 'aws_access_key_id={0};aws_secret_access_key={1}'.format(
+            access_key_id,
+            secret_access_key
+        )
+
+        if session_token is not None:
+            credentials += ';token={0}'.format(session_token)
+
+        if not creds_rx.match(credentials):
+            raise ValueError('credentials must match the following'
+                             ' format:\n'
+                             'aws_access_key_id=<access-key-id>;'
+                             'aws_secret_access_key=<secret-access-key>'
+                             '[;token=<temporary-session-token>]\ngot %r' %
+                             credentials)
+
+        if len(delimiter) != 1:
+            raise ValueError('"delimiter" parameter must be a single '
+                             'character')
+
+        if not isinstance(ignore_header, numbers.Integral):
+            raise TypeError('"ignore_header" parameter should be an integer')
+
+        if format not in self.formats:
+            raise ValueError('"format" parameter must be one of %s' %
+                             self.formats)
+
+        if compression is not None and compression not in self.compression_types:
+            raise ValueError('"compression" parameter must be one of %s' %
+                             self.compression_types)
+
+        self.table = table
         self.data_location = data_location
-        self.access_key = access_key
-        self.secret_key = secret_key
-        self.session_token = session_token
-        self.options = options
+        self.credentials = credentials
+        self.delimiter = delimiter
+        self.ignore_header = ignore_header
+        self.dangerous_null_delimiter = dangerous_null_delimiter
+        self.manifest = manifest
+        self.empty_as_null = empty_as_null
+        self.blanks_as_null = blanks_as_null
+        self.format = format
+        self.compression = compression or ''
+
+
+def _tablename(t, compiler):
+    name = compiler.preparer.quote(t.name)
+    if t.schema is not None:
+        return '%s.%s' % (compiler.preparer.quote_schema(t.schema), name)
+    else:
+        return name
 
 
 @compiles(CopyCommand)
 def visit_copy_command(element, compiler, **kw):
     ''' Returns the actual sql query for the CopyCommand class
     '''
-    return """
-           COPY %(schema_name)s.%(table_name)s FROM '%(data_location)s'
-           CREDENTIALS 'aws_access_key_id=%(access_key)s;aws_secret_access_key=%(secret_key)s%(session_token)s'
-           CSV
-           TRUNCATECOLUMNS
-           DELIMITER '%(delimiter)s'
-           IGNOREHEADER %(ignore_header)s
-           %(null)s
-           %(manifest)s
-           %(empty_as_null)s
-           %(blanks_as_null)s;
-           """ % \
-           {'schema_name': element.schema_name,
-            'table_name': element.table_name,
-            'data_location': element.data_location,
-            'access_key': element.access_key,
-            'secret_key': element.secret_key,
-            'session_token': ';token=%s' % element.session_token if element.session_token else '',
-            'null': ("NULL '%s'" % element.options.get('null')) if element.options.get('null') else '',
-            'delimiter': element.options.get('delimiter', ','),
-            'ignore_header': element.options.get('ignore_header', 0),
-            'manifest': 'MANIFEST' if bool(element.options.get('manifest', False)) else '',
-            'empty_as_null': 'EMPTYASNULL' if bool(element.options.get('empty_as_null', True)) else '',
-            'blanks_as_null': 'BLANKSASNULL' if bool(element.options.get('blanks_as_null', True)) else ''}
+    qs = """COPY {table} FROM :data_location
+    CREDENTIALS :credentials
+    {format}
+    TRUNCATECOLUMNS
+    DELIMITER :delimiter
+    IGNOREHEADER :ignore_header
+    {null}
+    {manifest}
+    {compression}
+    {empty_as_null}
+    {blanks_as_null}
+    """.format(table=_tablename(element.table, compiler),
+               format=element.format,
+               manifest='MANIFEST' if element.manifest else '',
+               compression=element.compression,
+               empty_as_null='EMPTYASNULL' if element.empty_as_null else '',
+               blanks_as_null='BLANKSASNULL' if element.blanks_as_null else '',
+               ignore_header=element.ignore_header,
+               null=(("NULL '%s'" % element.dangerous_null_delimiter)
+                     if element.dangerous_null_delimiter is not None else ''))
+
+    return compiler.process(
+        sa.text(qs).bindparams(
+            sa.bindparam('data_location',
+                         value=element.data_location,
+                         type_=sa.String),
+            sa.bindparam('credentials', value=element.credentials,
+                         type_=sa.String),
+            sa.bindparam('delimiter',
+                         value=element.delimiter,
+                         type_=sa.String),
+            sa.bindparam('ignore_header',
+                         value=element.ignore_header,
+                         type_=sa.Integer)
+        ),
+        **kw
+    )
 
 
 @compiles(BindParameter)